mm: Reclaim early stack physical memory after bootstrap.
Now, early stack is a page reserved in the pageable memory pool,
and that page is freed at the end of the kernel bootstrap. In
such way that's as much as 2kB of physical memory that are
released to the system and the application for a later usage.

Early stack must be of the biggest page size available on the
system, such that no coalescing takes place.