syscall: Simplify branching in kernel
Now, SVC does not embed the number of arguments to the syscall,
but the syscall ID.

This patch removes the 'infinite syscall number' feature, but
simplifies a lot how syscall are managed in the kernel. It
also removes the possibility to plug new syscalls at runtime
but that should be straightforward to implement such functionnality
with the NVIC and Software Interrupt module.