Do kernel stack checking
The kernel stack layout is a Thread Control Block at the bottom of
the stack, and the stack storage from the top of the stack. The
stack grows toward the control block. There is a canary at the end
of the control block, such that the canary can be crunched in case
of a stack overflow.

It's not bullet-proof, still could prove itself helpful when
debugging hardfaults.