mutex: Yet another implementation of the feature
We have a global list of blocking threads, which needs to be
walked every time we unlock a mutex that has a non negative
value (i.e. it exists another thread in the system waiting for
that mutex). Linear complexity that could be improved in the
future (cache?), though it is very unlikely that there are many
different blocking mutexes at the same time in the system.