Remove the kernel stack canary feature
Kernel stack canary must be revamped. It should not be a field
in any structs because thread_info struct can be stored into
the task_struct with CONFIG_THREAD_INFO_IN_TASK. So actual end
of kernel stack can be either after the thread_info struct or
after the task_struct, depending on the config option.