arch/arm: Save/restore PSP on svcall entry/leave
Initially, the save/restore of SP_Process was done during the
call to __switch_to(). That was requiring a manual update of the
PSP value in the thread_info struct when the kernel needed to
access the process stack.

Now, the value is updated in the thread_info struct on exception
entry and leave. This comes at the cost of a few extra cycles.